Crate forest_actor_interface

Modules

account

cron

init

market

miner

multisig

power

reward

system

Macros

load_actor_state

Structs

AwardBlockRewardParams

FilterEstimate

Enums

ActorVersion

Map

Constants

EPOCHS_IN_DAY

EPOCH_DURATION_SECONDS

Duration of each tipset epoch.

Statics

BURNT_FUNDS_ACTOR_ADDR

CHAOS_ACTOR_ADDR

RESERVE_ADDRESS

Functions

actor_version

Returns an actor’s version or None if it was not a builtin

is_account_actor

Returns true if the code belongs to an account actor.

is_builtin_actor

Returns true if the code belongs to a builtin actor.

is_miner_actor

Returns true if the code belongs to a miner actor.

is_singleton_actor

Returns true if the code belongs to a singleton actor.