Crate forest_actor_interface

Modules

account

cron

init

market

miner

multisig

power

reward

system

Macros

load_actor_state

Structs

AwardBlockRewardParams

FilterEstimate

Enums

ActorVersion

Constants

BURNT_FUNDS_ACTOR_ADDR

CHAOS_ACTOR_ADDR

EPOCHS_IN_DAY

EPOCH_DURATION_SECONDS

Duration of each tipset epoch.

RESERVE_ADDRESS

Functions

is_account_actor

Returns true if the code belongs to an account actor.

is_miner_actor

Returns true if the code belongs to a miner actor.